refactor: AppUserService에서 DB 관련 코드 제거
